[Celebrity] Merge to code-quality
