+++ composer.json
@@ -64,7 +64,6 @@
             "Rector\\Autodiscovery\\": "rules/autodiscovery/src",
             "Rector\\BetterPhpDocParser\\": "packages/better-php-doc-parser/src",
             "Rector\\CakePHP\\": "rules/cakephp/src",
-            "Rector\\Celebrity\\": "rules/celebrity/src",
             "Rector\\ChangesReporting\\": "packages/changes-reporting/src",
             "Rector\\CodeQuality\\": "rules/code-quality/src",
             "Rector\\NodeRemoval\\": "packages/node-removal/src",
@@ -156,7 +155,6 @@
             "Rector\\Autodiscovery\\Tests\\": "rules/autodiscovery/tests",
             "Rector\\BetterPhpDocParser\\Tests\\": "packages/better-php-doc-parser/tests",
             "Rector\\CakePHP\\Tests\\": "rules/cakephp/tests",
-            "Rector\\Celebrity\\Tests\\": "rules/celebrity/tests",
             "Rector\\CodeQuality\\Tests\\": "rules/code-quality/tests",
             "Rector\\Caching\\Tests\\": "packages/caching/tests",
             "Rector\\CodingStyle\\Tests\\": "rules/coding-style/tests",
