+++ rules/code-quality/src/Rector/NotEqual/CommonNotEqualRector.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Celebrity\Rector\NotEqual;
+namespace Rector\CodeQuality\Rector\NotEqual;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr\BinaryOp\NotEqual;
@@ -13,7 +13,7 @@ use Rector\NodeTypeResolver\Node\AttributeKey;
 
 /**
  * @see https://stackoverflow.com/a/4294663/1348344
- * @see \Rector\Celebrity\Tests\Rector\NotEqual\CommonNotEqualRector\CommonNotEqualRectorTest
+ * @see \Rector\CodeQuality\Tests\Rector\NotEqual\CommonNotEqualRector\CommonNotEqualRectorTest
  */
 final class CommonNotEqualRector extends AbstractRector
 {
