+++ rules/code-quality/tests/Rector/BooleanOp/LogicalToBooleanRector/Fixture/and.php.inc
