+++ rules/code-quality/tests/Rector/BooleanOp/LogicalToBooleanRector/LogicalToBooleanRectorTest.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Rector\Celebrity\Tests\Rector\NotEqual\CommonNotEqualRector;
+namespace Rector\CodeQuality\Tests\Rector\BooleanOp\LogicalToBooleanRector;
 
 use Iterator;
-use Rector\Celebrity\Rector\NotEqual\CommonNotEqualRector;
+use Rector\CodeQuality\Rector\BooleanOp\LogicalToBooleanRector;
 use Rector\Core\Testing\PHPUnit\AbstractRectorTestCase;
 use Symplify\SmartFileSystem\SmartFileInfo;
 
-final class CommonNotEqualRectorTest extends AbstractRectorTestCase
+final class LogicalToBooleanRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideData()
@@ -26,6 +26,6 @@ final class CommonNotEqualRectorTest extends AbstractRectorTestCase
 
     protected function getRectorClass(): string
     {
-        return CommonNotEqualRector::class;
+        return LogicalToBooleanRector::class;
     }
 }
