+++ rules/code-quality/tests/Rector/FuncCall/SetTypeToCastRector/Fixture/fixture.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Celebrity\Tests\Rector\FuncCall\SetTypeToCastRector\Fixture;
+namespace Rector\CodeQuality\Tests\Rector\FuncCall\SetTypeToCastRector\Fixture;
 
 class SomeClass
 {
@@ -27,7 +27,7 @@ class SomeClass
 -----
 <?php
 
-namespace Rector\Celebrity\Tests\Rector\FuncCall\SetTypeToCastRector\Fixture;
+namespace Rector\CodeQuality\Tests\Rector\FuncCall\SetTypeToCastRector\Fixture;
 
 class SomeClass
 {
